tests/submission/W3C/domhtml/HTMLMetaElement02.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
<title>NIST DOM HTML Test - META</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-meta-element.html#dom-meta-httpequiv" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<br>
</p>



<script>
var domtest = async_test("The httpEquiv attribute specifies an HTTP respnse header name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhttpequiv;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("meta");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vhttpequiv = testNode.httpEquiv;

      assert_equals(vhttpequiv,"Content-Type");
       
    domtest.done();
}));

</script>
<p>The httpEquiv attribute specifies an HTTP respnse header name.

    Retrieve the httpEquiv attribute and examine its value.</p>
<div id="log"></div>
</body></html>
